Added Hardware window
Has been added a new section in configuration window. It displays the hardware information of the computer where it is executing the Engine.

Moreover, has been implemented some code to force the engine to be executed with NVIDIA Graphics card instead of the default one that the computer has configured.